// src/business/business/guidebook/types.ts
export enum EditorFieldEnum {
  // Arrival instructions
  welcomeMessage = 'Welcome Message',
  parkingInstructions = 'Parking Instructions',
  checkIn = 'Check-In Procedure',
  gettingToUnit = 'Directions to Property',
  frontDesk = 'Front-Desk Details',
  whatToBring = 'What to bring with you',
  lateCheckIn = 'Late Check-In Procedure',

  // Stay Instructions
  locationSummary = 'Location Summary',
  heating = 'Heating',
  aboutBuilding = 'About the Building',
  emergencyContactProcedure = 'Emergency Contact Procedure',
  directionsToSkiLift = 'Directions To Ski Lifts',
  skiInOut = 'Ski In / Ski Out Instructions',
  gettingAround = 'Getting Around the Area',
  linens = 'Extra Towels & Linens',
  restaurants = 'Restaurants',
  services = 'Services & Activities',
  houseKeeping = 'House Keeping Details',
  other = 'Other',

  // Departure Instructions
  checkOut = 'Check-Out Process',
  garbage = 'Garbage Removal',
}